requirements.txt
langgraph>=0.2.50
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.1
